+++ ado_wrapper/resources/runs.py
@@ -15,8 +15,8 @@ RunState = Literal["canceling", "completed", "inProgress", "unknown"]
 
 
 class RunAllDictionary(TypedDict):
-    template_parameters: NotRequired[dict[str, Any]]
-    run_variables: NotRequired[dict[str, Any]]
+    template_parameters: NotRequired[dict[str, Any] | None]
+    run_variables: NotRequired[dict[str, Any] | None]
     branch_name: NotRequired[str]
     stages_to_run: NotRequired[list[str] | None]
 
@@ -52,31 +52,32 @@ class Run(StateManagedResource):
 
     @classmethod
     def create(
-        cls, ado_client: "AdoClient", definition_id: str, template_parameters: dict[str, Any] = {}, run_variables: dict[str, Any] = {},
-        source_branch: str = "main", stages_to_run: list[str] | None = None  # fmt: skip
+        cls, ado_client: "AdoClient", definition_id: str, template_parameters: dict[str, Any] | None = None,
+        run_variables: dict[str, Any] | None = None, branch_name: str = "main", stages_to_run: list[str] | None = None  # fmt: skip
     ) -> "Run":
-        """ Creates a `Run` in ADO and returns the object. If stages_to_run isn't set (or is set to None), all stages will be run. """
+        """Creates a `Run` in ADO and returns the object. If stages_to_run isn't set (or is set to None), all stages will be run."""
 
-        PAYLOAD = {"templateParameters": template_parameters, "variables": run_variables, "repositories": {"refName": f"refs/heads/{source_branch}"}}
+        PAYLOAD: dict[str, Any] = {
+            "templateParameters": template_parameters or {},
+            "variables": run_variables or {},
+            "repositories": {"refName": f"refs/heads/{branch_name}"},
+        }
         if stages_to_run is not None:
-            build_stages = BuildDefinition.get_all_stages(ado_client, definition_id, source_branch)
-        
+            build_stages = BuildDefinition.get_all_stages(ado_client, definition_id, branch_name)
+
             for name in stages_to_run:
                 if name not in {step.stage_internal_name for step in build_stages}:
                     raise ValueError(f"The stage_name '{name}' in stages_to_run is not found in the steps.")
 
-            PAYLOAD["stagesToSkip"] = [stage.stage_internal_name for stage in build_stages if stage.stage_internal_name not in stages_to_run]
+            PAYLOAD["stagesToSkip"] = [stage.stage_internal_name for stage in build_stages if stage.stage_internal_name not in stages_to_run]  # fmt: skip
         try:
-            print(f"/{ado_client.ado_project}/_apis/pipelines/{definition_id}/runs?api-version=6.1-preview.1")
             return super()._create(
                 ado_client,
                 f"/{ado_client.ado_project}/_apis/pipelines/{definition_id}/runs?api-version=6.1-preview.1",
                 PAYLOAD,
             )  # type: ignore[return-value]
         except ValueError as e:
-            raise ValueError(
-                f"A template variable inputted is not allowed! {str(e).split('message')[1][3:].removesuffix(':').split('.')[0]}"
-            ) from e
+            raise ValueError(f"A template parameter inputted is not allowed! {str(e).split('message')[1][3:].removesuffix(':').split('.')[0]}")  # fmt: skip
 
     @classmethod
     def delete_by_id(cls, ado_client: "AdoClient", run_id: str) -> None:
@@ -98,17 +99,17 @@ class Run(StateManagedResource):
 
     @classmethod
     def run_and_wait_until_completion(
-        cls, ado_client: "AdoClient", definition_id: str, 
-      
-      iables: dict[str, Any] = {}, run_variables: dict[str, Any] = {},
-        branch_name: str = "main", max_timeout_seconds: int | None = 900, stages_to_run: list[str] | None = None   # fmt: skip
+        cls, ado_client: "AdoClient", definition_id: str, template_parameters: dict[str, Any] | None = None, run_variables: dict[str, Any] | None = None,
+        branch_name: str = "main", stages_to_run: list[str] | None = None, max_timeout_seconds: int | None = 900  # fmt: skip
     ) -> "Run":
-        """Creates a run and waits until it is completed, or raises
-        a TimeoutError if it takes too long.
+        """Creates a run and waits until it is completed, or raises a TimeoutError if it takes too long.
         WARNING: This is a blocking operation, it will not return until the run is completed or the timeout (default 15 mins) is reached."""
-        data: dict[str, RunAllDictionary] = {definition_id: {
-          "template_parameters": template_parameters, "run_variables": run_variables, "branch_name": branch_name, "stages_to_run": stages_to_run  # fmt: skip
-        }}
+        data: dict[str, RunAllDictionary] = {
+            definition_id: {
+                "template_parameters": template_parameters, "run_variables": run_variables,
+                "branch_name": branch_name, "stages_to_run": stages_to_run
+            }  # fmt: skip
+        }
         return cls.run_all_and_capture_results_simultaneously(ado_client, data, max_timeout_seconds)[definition_id]
 
     @classmethod
@@ -118,9 +119,11 @@ class Run(StateManagedResource):
         """Takes a mapping of definition_id -> {template_parameters, run_variables, branch_name, stages_to_run}
         Once done, returns a mapping of definition_id -> `Run` object"""
         return_values = {}
-        for definition_id, run_data in data.items():            
-            run = cls.run_and_wait_until_completion(ado_client, definition_id, run_data["template_parameters"],
-                                                    run_data["run_variables"], run_data["branch_name"], run_data["stages_to_run"])  # fmt: skip
+        for definition_id, run_data in data.items():
+            run = cls.run_and_wait_until_completion(
+                ado_client, definition_id, run_data.get("template_parameters", {}), run_data.get("run_variables", {}),
+                run_data.get("branch_name", "main"), run_data.get("stages_to_run"), max_timeout_seconds  # fmt: skip
+            )
             return_values[definition_id] = run
         return return_values
 
@@ -133,8 +136,10 @@ class Run(StateManagedResource):
         # Get a mapping of definition_id -> Run()
         runs: dict[str, Run] = {}
         for definition_id, run_data in data.items():
-            run = cls.create(ado_client, definition_id, run_data["template_parameters"],
-                             run_data["run_variables"], run_data["branch_name"], run_data["stages_to_run"])
+            run = cls.create(
+                ado_client, definition_id, run_data.get("template_parameters", {}), run_data.get("run_variables", {}),
+                run_data.get("branch_name", "main"), run_data.get("stages_to_run"),  # fmt: skip
+            )
             runs[definition_id] = run
         # Then, slowly check on them, and remove the ones that are done
         start_time = datetime.now()